improve cyclomatic complexity
