+++ internal/synchronizer/client_test.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"io"
 	"log"
+	"nabu/internal/common"
 	"nabu/internal/synchronizer/s3"
 	"nabu/internal/synchronizer/triplestore"
+	"nabu/pkg/config"
 	testhelpers "nabu/testHelpers"
 	"net/http"
 	"os"
@@ -20,6 +22,24 @@ import (
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/network"
 )
+
+// Create a new SynchronizerClient by directly passing in the clients
+// Mainly used for testing
+func newSynchronizerClient(graphClient *triplestore.GraphDbClient, s3Client *s3.MinioClientWrapper, bucketName string) (SynchronizerClient, error) {
+	processor, options, err := common.NewJsonldProcessor(config.NabuConfig{})
+	if err != nil {
+		return SynchronizerClient{}, err
+	}
+
+	client := SynchronizerClient{
+		GraphClient:     graphClient,
+		S3Client:        s3Client,
+		syncBucketName:  bucketName,
+		jsonldProcessor: processor,
+		jsonldOptions:   options,
+	}
+	return client, nil
+}
 
 func countSourcesInSitemap(url string) (int, error) {
 	// Fetch the URL
@@ -121,13 +141,13 @@ func (suite *SynchronizerClientSuite) TestMoveObjToTriplestore() {
 	summonedObjs, err := suite.client.S3Client.NumberOfMatchingObjects([]string{"summoned/cdss0/"})
 	require.NoError(t, err)
 	require.Equal(t, sourcesInSitemap, summonedObjs)
-	err = suite.client.CopyAllPrefixedObjToTriplestore([]string{"orgs/"})
+	err = suite.client.CopyAllPrefixedObjToTriplestore("orgs/")
 	require.NoError(t, err)
 	graphs, err := suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("orgs/")
 	require.NoError(t, err)
 	require.Len(t, graphs, 1)
 
-	err = suite.client.CopyAllPrefixedObjToTriplestore([]string{"summoned/"})
+	err = suite.client.CopyAllPrefixedObjToTriplestore("summoned/")
 	require.NoError(t, err)
 	graphs, err = suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("summoned/")
 	require.NoError(t, err)
@@ -178,7 +198,7 @@ func (suite *SynchronizerClientSuite) TestSyncTriplestore() {
 
 	// make sure that an old graph is no longer there when
 	// we sync new org data
-	err = suite.client.SyncTriplestoreGraphs([]string{"orgs/"})
+	err = suite.client.SyncTriplestoreGraphs("orgs/")
 	require.NoError(t, err)
 	exists, err = suite.graphdbContainer.Client.GraphExists(oldGraph)
 	require.False(t, exists)
@@ -189,7 +209,7 @@ func (suite *SynchronizerClientSuite) TestSyncTriplestore() {
 	require.Equal(t, len(graphs), 1)
 
 	// make sure that there is prov data for every source in the sitemap
-	err = suite.client.SyncTriplestoreGraphs([]string{"prov/"})
+	err = suite.client.SyncTriplestoreGraphs("prov/")
 	require.NoError(t, err)
 	graphs, err = suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("prov/")
 	require.NoError(t, err)
@@ -201,7 +221,7 @@ func (suite *SynchronizerClientSuite) TestSyncTriplestore() {
 	require.Equal(t, len(graphs), 1)
 
 	// make sure that summoned data matches the amount of sources in the sitemap
-	err = suite.client.SyncTriplestoreGraphs([]string{"summoned/"})
+	err = suite.client.SyncTriplestoreGraphs("summoned/")
 	require.NoError(t, err)
 	graphs, err = suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("summoned/")
 	require.NoError(t, err)
@@ -222,7 +242,7 @@ func (suite *SynchronizerClientSuite) TestSyncTriplestore() {
 
 	// make sure that graph syncs are additive between sources and that
 	// sources are not overwritten or removed
-	err = suite.client.SyncTriplestoreGraphs([]string{"summoned/refgages0"})
+	err = suite.client.SyncTriplestoreGraphs("summoned/refgages0")
 	require.NoError(t, err)
 	graphs, err = suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("summoned/")
 	require.NoError(t, err)
@@ -234,7 +254,7 @@ func (suite *SynchronizerClientSuite) TestSyncTriplestore() {
 	require.NoError(t, err)
 	err = suite.client.S3Client.Remove(objs[0].Key)
 	require.NoError(t, err)
-	err = suite.client.SyncTriplestoreGraphs([]string{"summoned/"})
+	err = suite.client.SyncTriplestoreGraphs("summoned/")
 	require.NoError(t, err)
 	graphs, err = suite.client.GraphClient.NamedGraphsAssociatedWithS3Prefix("summoned/")
 	require.NoError(t, err)
@@ -256,7 +276,7 @@ func (suite *SynchronizerClientSuite) TestNqRelease() {
 	require.NoError(t, err)
 	require.Zero(t, gleanerContainer.ExitCode, gleanerContainer.Logs)
 
-	err = suite.client.GenerateNqRelease([]string{"orgs/cdss0"})
+	err = suite.client.GenerateNqRelease("orgs/cdss0")
 	require.NoError(t, err)
 	const nqPath = "graphs/latest/cdss0_organizations.nq"
 	objs, err := suite.client.S3Client.NumberOfMatchingObjects([]string{nqPath})
