+++ pkg/cli/release.go
@@ -14,7 +14,12 @@ func release() error {
 		return err
 	}
 
-	err = client.GenerateNqRelease(cfgStruct.Prefixes)
+	for _, prefix := range cfgStruct.Prefixes {
+		err = client.CopyAllPrefixedObjToTriplestore(prefix)
+		if err != nil {
+			return err
+		}
+	}
 
 	return err
 }
